Contract interface is actually a runtime
