+++ crates/locutus-node/src/contract/handler.rs
@@ -1,8 +1,9 @@
-use crate::contract::interface::{ContractInterface, ContractUpdateResult};
 use crate::contract::store::ContractStore;
 use crate::contract::{ContractHandler, ContractHandlerChannel, ContractKey, ContractValue};
 use sqlx::sqlite::SqliteRow;
 use sqlx::{Row, SqlitePool};
+
+use super::runtime::ContractRuntime;
 
 #[derive(Debug, thiserror::Error)]
 pub(crate) enum DatabaseError {
@@ -10,25 +11,6 @@ pub(crate) enum DatabaseError {
     ContractNotFound,
     #[error(transparent)]
     SqliteError(#[from] sqlx::Error),
-}
-
-pub(crate) struct TestContract {}
-
-impl ContractInterface for TestContract {
-    fn validate_value(value: &[u8]) -> bool {
-        todo!()
-    }
-
-    fn update_value(value: Vec<u8>, value_update: &[u8]) -> ContractUpdateResult<Vec<u8>> {
-        Ok(value_update.to_vec())
-    }
-    fn related_contracts(value_update: &[u8]) -> Vec<ContractKey> {
-        todo!()
-    }
-
-    fn extract(extractor: Option<&[u8]>, value: &[u8]) -> Vec<u8> {
-        todo!()
-    }
 }
 
 pub(crate) struct SQLiteContractHandler {
@@ -97,7 +79,7 @@ impl ContractHandler for SQLiteContractHandler {
             Err(_) => value.clone(),
         };
 
-        let value: Vec<u8> = match TestContract::update_value(old_value.0, &value.0) {
+        let value: Vec<u8> = match ContractRuntime::update_value(old_value.0, &value.0) {
             Ok(contract_value) => contract_value,
             Err(err) => err.value,
         };
@@ -118,5 +100,75 @@ impl ContractHandler for SQLiteContractHandler {
             Ok(contract_value) => Ok(contract_value),
             Err(err) => Err(DatabaseError::SqliteError(err)),
         }
+    }
+}
+#[cfg(test)]
+mod test {
+    use crate::contract::Contract;
+
+    use super::*;
+
+    // Prepare and get handler for an in-memory sqlite db
+    async fn get_handler() -> Result<SQLiteContractHandler, sqlx::Error> {
+        let ch_handler = ContractHandlerChannel::new();
+        let db_pool = SqlitePool::connect("sqlite::memory:").await.unwrap();
+        let store: ContractStore = ContractStore::new();
+        create_test_contracts_table(&db_pool).await;
+        Ok(SQLiteContractHandler::new(ch_handler, store, db_pool))
+    }
+
+    // Create test contracts table
+    async fn create_test_contracts_table(pool: &SqlitePool) {
+        sqlx::query(
+            "CREATE TABLE IF NOT EXISTS contracts (
+        key             STRING PRIMARY KEY,
+        value           BLOB
+        )",
+        )
+        .execute(pool)
+        .await
+        .unwrap();
+    }
+
+    #[tokio::test]
+    async fn contract_handler() -> Result<(), anyhow::Error> {
+        // Create a sqlite handler and initialize the database
+        let mut handler: SQLiteContractHandler = get_handler().await?;
+        create_test_contracts_table(&handler.pool).await;
+
+        // Generate a contract
+        let contract_value: Vec<u8> = b"Test contract value".to_vec();
+        let contract: Contract = Contract::new(contract_value);
+
+        // Get contract parts
+        let contract_key = ContractKey(contract.key);
+        let contract_value = ContractValue::new(contract.data);
+        let contract_value_cloned = contract_value.clone();
+
+        let put_result_value = handler.put_value(&contract_key, contract_value).await?;
+        let get_result_value = handler
+            .get_value(&contract_key)
+            .await?
+            .ok_or(anyhow::anyhow!("No value found"))?;
+
+        assert_eq!(contract_value_cloned.0, put_result_value.0);
+        assert_eq!(contract_value_cloned.0, get_result_value.0);
+        assert_eq!(put_result_value.0, get_result_value.0);
+
+        // Update the contract value with new one
+        let new_contract_value: Vec<u8> = b"New test contract value".to_vec();
+        let new_contract_value = ContractValue::new(new_contract_value);
+        let new_contract_value_cloned = new_contract_value.clone();
+        let new_put_result_value = handler.put_value(&contract_key, new_contract_value).await?;
+        let new_get_result_value = handler
+            .get_value(&contract_key)
+            .await?
+            .ok_or(anyhow::anyhow!("No value found"))?;
+
+        assert_eq!(new_contract_value_cloned.0, new_put_result_value.0);
+        assert_eq!(new_contract_value_cloned.0, new_get_result_value.0);
+        assert_eq!(new_put_result_value.0, new_get_result_value.0);
+
+        Ok(())
     }
 }
