+++ crates/locutus-node/src/contract/mod.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Deserializer, Serialize};
 use crate::ring::Location;
 
 mod handler;
-mod interface;
+mod runtime;
 mod store;
 mod test_utils;
