+++ crates/locutus-node/src/contract/runtime.rs
@@ -3,22 +3,17 @@
 //! All valid contracts must implement this interface.
 //!
 //! This abstraction layer shouldn't leak beyond the contract handler.
-
 use super::ContractKey;
 
 pub(super) type ContractUpdateResult<T> = Result<T, ContractUpdateError>;
 
-pub(super) struct ContractUpdateError {
-    /// original PUT value
-    pub value: Vec<u8>,
-    kind: ErrorKind,
-}
+pub(crate) struct ContractRuntime {}
 
-pub(super) enum ErrorKind {}
-
-pub(super) trait ContractInterface {
+impl ContractRuntime {
     /// Determine whether this value is valid for this contract
-    fn validate_value(value: &[u8]) -> bool;
+    pub fn validate_value(value: &[u8]) -> bool {
+        todo!()
+    }
 
     /// Determine whether this value is a valid update for this contract. If it is, return the modified value,
     /// else return error and the original value.
@@ -27,16 +22,30 @@ pub(super) trait ContractInterface {
     /// - If the same `value_update` is applied twice to a value, then the second will be ignored.
     /// - Application of `value_update` is "order invariant", no matter what the order in which the values are
     ///   applied, the resulting value must be exactly the same.
-    fn update_value(value: Vec<u8>, value_update: &[u8]) -> ContractUpdateResult<Vec<u8>>;
+    pub fn update_value(value: Vec<u8>, value_update: &[u8]) -> ContractUpdateResult<Vec<u8>> {
+        Ok(value_update.to_vec())
+    }
 
     /// Obtain any other related contracts for this value update. Typically used to ensure
     /// update has been fully propagated.
-    fn related_contracts(value_update: &[u8]) -> Vec<ContractKey>;
+    pub fn related_contracts(value_update: &[u8]) -> Vec<ContractKey> {
+        todo!()
+    }
 
     /// Extract some data from the value and return it.
     /// E.g. `extractor` might contain a byte range, which will be extracted
     /// from the value and returned.
     ///
     /// In case extractor is none, then return the whole value.
-    fn extract(extractor: Option<&[u8]>, value: &[u8]) -> Vec<u8>;
+    pub fn extract(extractor: Option<&[u8]>, value: &[u8]) -> Vec<u8> {
+        todo!()
+    }
 }
+
+pub(crate) struct ContractUpdateError {
+    /// original PUT value
+    pub value: Vec<u8>,
+    pub kind: ErrorKind,
+}
+
+pub(crate) enum ErrorKind {}
